Review fixes

+++ packaging/rpm/create-rpm-repo.sh
@@ -79,6 +79,10 @@ for arch in "${ARCHES[@]}"; do
 
   for rpm in "${rpms[@]}"; do
     target="$REPODIR/$arch/$(basename "$rpm")"
+    if [[ -f "$target" ]]; then
+      echo "Skipping existing: $(basename "$rpm")"
+      continue
+    fi
     echo "Copying: $(basename "$rpm")"
     cp "$rpm" "$target"
   done
@@ -105,6 +109,10 @@ mkdir -p "$REPODIR/SRPMS"
 
 for srpm in "${srpms[@]}"; do
   target="$REPODIR/SRPMS/$(basename "$srpm")"
+  if [[ -f "$target" ]]; then
+    echo "Skipping existing: $(basename "$rpm")"
+    continue
+  fi
   echo "Copying: $(basename "$srpm")"
   cp "$srpm" "$target"
 done
